feat: develop 머지 및 코드 수정
