+++ cookie/src/main/java/com/cookie/admin/service/AdminMovieCreateService.java
@@ -4,12 +4,16 @@ import com.cookie.admin.dto.response.AdminMovieBaseAddResponse;
 import com.cookie.admin.dto.response.AdminMovieDetailResponse;
 import com.cookie.admin.dto.response.tmdb.TMDBMovieSearchResponse;
 import com.cookie.admin.exception.MovieAlreadyExistsException;
-import com.cookie.admin.repository.*;
+import com.cookie.admin.repository.CategoryRepository;
+import com.cookie.admin.repository.CountryRepository;
 import com.cookie.domain.actor.entity.Actor;
+import com.cookie.domain.actor.repository.ActorRepository;
 import com.cookie.domain.category.entity.Category;
 import com.cookie.domain.country.entity.Country;
 import com.cookie.domain.director.entity.Director;
+import com.cookie.domain.director.repository.DirectorRepository;
 import com.cookie.domain.movie.entity.*;
+import com.cookie.domain.movie.repository.*;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -21,7 +25,6 @@ import reactor.core.publisher.Mono;
 import reactor.util.retry.Retry;
 
 import java.time.Duration;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
