+++ cookie/src/main/java/com/cookie/admin/service/AdminMovieModifyService.java
@@ -4,10 +4,11 @@ import com.cookie.admin.dto.response.AdminMovieCategoryResponse;
 import com.cookie.admin.dto.response.AdminMovieDeleteResponse;
 import com.cookie.admin.dto.response.MovieCategories;
 import com.cookie.admin.exception.MovieNotFoundException;
-import com.cookie.admin.repository.*;
+import com.cookie.admin.repository.CategoryRepository;
 import com.cookie.domain.category.entity.Category;
 import com.cookie.domain.movie.entity.Movie;
 import com.cookie.domain.movie.entity.MovieCategory;
+import com.cookie.domain.movie.repository.*;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
