+++ cookie/src/main/java/com/cookie/domain/movie/dto/response/MovieResponse.java
@@ -16,11 +16,10 @@ public class MovieResponse {
     private String title;              // 제목
     private String poster;             // 포스터 이미지
     private String plot;               // 줄거리
-    private String company;            // 제작사
     private String releasedAt;         // 개봉일 (LocalDateTime -> String 변환)
-    private String runtime;            // 상영시간 (int -> String 변환)
+    private Integer runtime;           // 상영시간 (Integer)
     private Double score;              // 평점
-    private String rating;             // 연령등급 (Enum -> String 변환)
+    private String certification;      // 연령등급 (Enum -> String 변환)
     private List<String> images;       // 영화 이미지 리스트
     private List<MovieVideoResponse> videos; // 영화 비디오 리스트
     private List<String> countries;    // 제작 국가 리스트
