+++ cookie/src/main/java/com/cookie/domain/movie/dto/response/ReviewOfMovieResponse.java
@@ -1,11 +1,9 @@
 package com.cookie.domain.movie.dto.response;
 
-import com.cookie.domain.review.dto.response.MovieReviewResponse;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
@@ -14,10 +12,10 @@ import java.util.List;
 public class ReviewOfMovieResponse {
     private String title;
     private String poster;
-    private String rating;
+    private String certification;
     private int runtime;
     private List<String> subCategories;
     private List<String> countries;
-    private LocalDateTime releasedAt;
+    private String releasedAt;
     private List<MovieReviewResponse> reviews;
 }
