+++ cookie/src/main/java/com/cookie/domain/movie/repository/MovieCountryRepository.java
@@ -2,7 +2,9 @@ package com.cookie.domain.movie.repository;
 
 import com.cookie.domain.movie.entity.MovieCountry;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -11,4 +13,11 @@ import java.util.List;
 public interface MovieCountryRepository extends JpaRepository<MovieCountry, Long> {
     @Query("SELECT mc FROM MovieCountry mc JOIN FETCH mc.country c WHERE mc.movie.id = :movieId")
     List<MovieCountry> findByMovieIdWithCountry(Long movieId);
+
+    @Modifying
+    @Query("""
+        DELETE FROM MovieCountry  ma
+        WHERE ma.movie.id = :movieId
+    """)
+    void deleteByMovieId(@Param("movieId") Long movieId);
 }
