+++ cookie/src/main/java/com/cookie/domain/movie/repository/MovieImageRepository.java
@@ -1,16 +1,16 @@
-package com.cookie.admin.repository;
+package com.cookie.domain.movie.repository;
 
-import com.cookie.domain.movie.entity.MovieCountry;
+import com.cookie.domain.movie.entity.MovieImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface MovieCountryRepository extends JpaRepository<MovieCountry, Long> {
+public interface MovieImageRepository extends JpaRepository<MovieImage, Long> {
 
     @Modifying
     @Query("""
-        DELETE FROM MovieCountry  ma
+        DELETE FROM MovieImage  ma
         WHERE ma.movie.id = :movieId
     """)
     void deleteByMovieId(@Param("movieId") Long movieId);
