+++ cookie/src/main/java/com/cookie/domain/movie/repository/MovieRepository.java
@@ -1,8 +1,12 @@
 package com.cookie.domain.movie.repository;
 
 import com.cookie.domain.movie.entity.Movie;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
@@ -24,4 +28,25 @@ public interface MovieRepository extends JpaRepository<Movie, Long> {
     // 영화 국가 가져오기
     @Query("SELECT m FROM Movie m LEFT JOIN FETCH m.movieCountries mc LEFT JOIN FETCH mc.country WHERE m.id = :movieId")
     Optional<Movie> findByIdWithCountries(Long movieId);
+
+    @Query("""
+        SELECT m
+        FROM Movie m
+        WHERE m.TMDBMovieId = :TMDBMovieId
+    """)
+    Optional<Movie> findMovieByTMDBMovieId(@Param("TMDBMovieId") Long TMDBMovieId);
+
+    @Query("""
+        SELECT m
+        FROM Movie m
+        WHERE m.title LIKE CONCAT('%', :title, '%')
+    """)
+    Page<Movie> findMovieByTitle(@Param("title") String title, Pageable page);
+
+    @Modifying
+    @Query("""
+        DELETE FROM Movie  m
+        WHERE m.id = :movieId
+    """)
+    void deleteByMovieId(@Param("movieId") Long movieId);
 }
