+++ cookie/src/main/java/com/cookie/domain/movie/repository/MovieVideoRepository.java
@@ -1,17 +1,17 @@
-package com.cookie.admin.repository;
+package com.cookie.domain.movie.repository;
 
-import com.cookie.domain.movie.entity.MovieImage;
+import com.cookie.domain.movie.entity.MovieVideo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface MovieImageRepository extends JpaRepository<MovieImage, Long> {
+public interface MovieVideoRepository extends JpaRepository<MovieVideo, Long> {
 
     @Modifying
     @Query("""
-        DELETE FROM MovieImage  ma
-        WHERE ma.movie.id = :movieId
+        DELETE FROM MovieVideo  m
+        WHERE m.id = :movieId
     """)
     void deleteByMovieId(@Param("movieId") Long movieId);
 }
