--- cookie/src/main/java/com/cookie/domain/review/dto/response/MovieReviewResponse.java
@@ -1,22 +1,0 @@
-package com.cookie.domain.review.dto.response;
-
-import com.cookie.domain.user.dto.response.MovieReviewUserResponse;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MovieReviewResponse {
-
-    private String content;
-    private long reviewLike;
-    private double movieScore;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private MovieReviewUserResponse user;
-
-}
