--- cookie/src/main/java/com/cookie/domain/review/dto/response/ReviewCommentResponse.java
@@ -1,19 +1,0 @@
-package com.cookie.domain.review.dto.response;
-
-
-import com.cookie.domain.user.dto.response.CommentUserResponse;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ReviewCommentResponse {
-    private CommentUserResponse user;
-    private LocalDateTime createdAt;
-    private String comment;
-  
-}
