+++ cookie/src/main/java/com/cookie/domain/user/dto/response/MovieReviewUserResponse.java
@@ -1,6 +1,6 @@
 package com.cookie.domain.user.dto.response;
 
-import com.cookie.domain.review.dto.response.MovieReviewResponse;
+//import com.cookie.domain.review.dto.response.MovieReviewResponse;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
